+++ jest.config.js
@@ -8,10 +8,7 @@ module.exports = {
         '<rootDir>/node_modules/regenerator-runtime/runtime',
         '<rootDir>/config/polyfills.js',
     ],
-    setupFilesAfterEnv: [
-        '<rootDir>config/jest/setup.js',
-        '@testing-library/react/cleanup-after-each',
-    ],
+    setupFilesAfterEnv: ['<rootDir>config/jest/setup.js'],
     testMatch: [
         '<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs,ts,tsx}',
         '<rootDir>/src/**/*.(spec|test).{js,jsx,mjs,ts,tsx}',